feat: refactor swap panel components for improved structure and readability
